Conflicts:
	modules/openlmis-web/src/main/webapp/public/js/shared/services/reports.js

+++ modules/core/src/main/java/org/openlmis/core/repository/mapper/ProductMapper.java
@@ -62,7 +62,7 @@ public interface ProductMapper {
     "#{packRoundingThreshold}, #{category.id},  #{productGroup.id}," +
     "#{createdBy}, #{modifiedBy}, #{modifiedDate})")
   @Options(useGeneratedKeys = true)
-  Integer insert(Product product);
+  Long insert(Product product);
 
   @Select("SELECT id FROM dosage_Units WHERE LOWER(code) = LOWER(#{code})")
   Long getDosageUnitIdForCode(String code);
@@ -96,7 +96,7 @@ public interface ProductMapper {
     "modifiedBy=#{modifiedBy}, modifiedDate=#{modifiedDate} WHERE id=#{id}"})
   void update(Product product);
 
-  @Select("SELECT * FROM products WHERE id=#{id}")
+   @Select("SELECT * FROM products WHERE id=#{id}")
     @Results({
       @Result(property = "productGroup", column = "productGroupId", javaType = ProductGroup.class,
         one = @One(select = "org.openlmis.core.repository.mapper.ProductGroupMapper.getById"))
